image reading and encoding
