+++ CppCurl/CppCurl.cpp
@@ -1,6 +1,5 @@
 #include "CppCurl.h"
 #include <iostream>
-#include <cassert>
 #include <curl/curl.h>
 #pragma comment(lib,"libcurl.lib")
 
@@ -82,12 +81,12 @@ std::string formulaRecognition(const std::string& encoded_image, const std::stri
 
 void downloadRenderedFormula(const std::string& latex_string, const std::string& file_path, const std::string& format)
 {
-	std::string url = latex_rendering_request_url + format+ ".download?"+ latex_string;
+	std::string url = latex_rendering_request_url + format + ".download?" + latex_string;
 	CURL* curl = NULL;
 	CURLcode result_code;
 
 	FILE* fp = NULL;
-	fopen_s(&fp,file_path.c_str(),"wb");
+	fopen_s(&fp, file_path.c_str(), "wb");
 
 	curl = curl_easy_init();
 
@@ -98,9 +97,9 @@ void downloadRenderedFormula(const std::string& latex_string, const std::string&
 		// URL地址，需要char*
 		curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
 		// 设置写函数
-		curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION,callbackFormulaRender );
+		curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, callbackFormulaRender);
 		// 设置写函数写入的指针
-		curl_easy_setopt(curl,CURLOPT_WRITEDATA,fp);
+		curl_easy_setopt(curl, CURLOPT_WRITEDATA, fp);
 
 		result_code = curl_easy_perform(curl);
 		if (result_code != CURLE_OK)
@@ -117,163 +116,4 @@ void downloadRenderedFormula(const std::string& latex_string, const std::string&
 	{
 		throw std::runtime_error("curl_easy_init() failed.");
 	}
-}
-
-
-static const std::string base64_chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-
-
-static inline bool isBase64(const char c)
-{
-    return (isalnum(c) || (c == '+') || (c == '/'));
-}
-
-std::string base64Encode(const char* bytes_to_encode, unsigned int in_len)
-{
-	std::string ret;
-	int i = 0;
-	int j = 0;
-	unsigned char char_array_3[3];
-	unsigned char char_array_4[4];
-
-	while (in_len--)
-	{
-		char_array_3[i++] = *(bytes_to_encode++);
-		if (i == 3)
-		{
-			char_array_4[0] = (char_array_3[0] & 0xfc) >> 2;
-			char_array_4[1] = ((char_array_3[0] & 0x03) << 4) + ((char_array_3[1] & 0xf0) >> 4);
-			char_array_4[2] = ((char_array_3[1] & 0x0f) << 2) + ((char_array_3[2] & 0xc0) >> 6);
-			char_array_4[3] = char_array_3[2] & 0x3f;
-			for (i = 0; (i < 4); i++)
-			{
-				ret += base64_chars[char_array_4[i]];
-			}
-			i = 0;
-		}
-	}
-	if (i)
-	{
-		for (j = i; j < 3; j++)
-		{
-			char_array_3[j] = '\0';
-		}
-
-		char_array_4[0] = (char_array_3[0] & 0xfc) >> 2;
-		char_array_4[1] = ((char_array_3[0] & 0x03) << 4) + ((char_array_3[1] & 0xf0) >> 4);
-		char_array_4[2] = ((char_array_3[1] & 0x0f) << 2) + ((char_array_3[2] & 0xc0) >> 6);
-		char_array_4[3] = char_array_3[2] & 0x3f;
-
-		for (j = 0; (j < i + 1); j++)
-		{
-			ret += base64_chars[char_array_4[j]];
-		}
-
-		while ((i++ < 3))
-		{
-			ret += '=';
-		}
-
-	}
-	return ret;
-}
-
-std::string base64Decode(const std::string & encoded_string)
-{
-	int in_len = encoded_string.size();
-	int i = 0;
-	int j = 0;
-	int in_ = 0;
-	unsigned char char_array_4[4], char_array_3[3];
-	std::string ret;
-
-	while (in_len-- && (encoded_string[in_] != '=') && isBase64(encoded_string[in_])) {
-		char_array_4[i++] = encoded_string[in_]; in_++;
-		if (i == 4) {
-			for (i = 0; i < 4; i++)
-				char_array_4[i] = base64_chars.find(char_array_4[i]);
-
-			char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
-			char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
-			char_array_3[2] = ((char_array_4[2] & 0x3) << 6) + char_array_4[3];
-
-			for (i = 0; (i < 3); i++)
-				ret += char_array_3[i];
-			i = 0;
-		}
-	}
-	if (i) {
-		for (j = i; j < 4; j++)
-			char_array_4[j] = 0;
-
-		for (j = 0; j < 4; j++)
-			char_array_4[j] = base64_chars.find(char_array_4[j]);
-
-		char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
-		char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
-		char_array_3[2] = ((char_array_4[2] & 0x3) << 6) + char_array_4[3];
-
-		for (j = 0; (j < i - 1); j++) ret += char_array_3[j];
-	}
-
-	return ret;
-}
-
-
-unsigned char decToHex(unsigned char x)
-{
-	return  x > 9 ? x + 55 : x + 48;
-}
-
-unsigned char hexToDec(unsigned char x)
-{
-	unsigned char y;
-	if (x >= 'A' && x <= 'Z') y = x - 'A' + 10;
-	else if (x >= 'a' && x <= 'z') y = x - 'a' + 10;
-	else if (x >= '0' && x <= '9') y = x - '0';
-	else assert(0);
-	return y;
-}
-
-std::string urlEncode(const std::string& bytes_to_encode)
-{
-	std::string strTemp = "";
-	size_t length = bytes_to_encode.length();
-	for (size_t i = 0; i < length; i++)
-	{
-		if (isalnum(bytes_to_encode[i]) ||
-			(bytes_to_encode[i] == '-') ||
-			(bytes_to_encode[i] == '_') ||
-			(bytes_to_encode[i] == '.') ||
-			(bytes_to_encode[i] == '~'))
-			strTemp += bytes_to_encode[i];
-		else if (bytes_to_encode[i] == ' ')
-			strTemp += "+";
-		else
-		{
-			strTemp += '%';
-			strTemp += decToHex(bytes_to_encode[i] >> 4);
-			strTemp += decToHex(bytes_to_encode[i] % 16);
-		}
-	}
-	return strTemp;
-}
-
-std::string urlDecode(const std::string& encoded_string)
-{
-	std::string strTemp = "";
-	size_t length = encoded_string.length();
-	for (size_t i = 0; i < length; i++)
-	{
-		if (encoded_string[i] == '+') strTemp += ' ';
-		else if (encoded_string[i] == '%')
-		{
-			assert(i + 2 < length);
-			unsigned char high = hexToDec(encoded_string[++i]);
-			unsigned char low = hexToDec(encoded_string[++i]);
-			strTemp += static_cast<unsigned char>(high * 16 + low);
-		}
-		else strTemp += encoded_string[i];
-	}
-	return strTemp;
 }
