+++ CppCurl/CppCurl.h
@@ -13,28 +13,3 @@ std::string formulaRecognition(const std::string& encoded_image, const std::stri
 // file_path: 下载渲染后图片的路径.
 // format: 下载图片的格式. 支持gif, svg, png.
 void downloadRenderedFormula(const std::string& latex_string, const std::string& file_path, const std::string& format);
-
-// @param:
-// bytes_to_encode: 需要被编码的数据的指针.
-// in_len: 需要被编码的数据的长度.
-// @return:
-// base64编码后的数据.
-std::string base64Encode(const char* bytes_to_encode, unsigned int in_len);
-
-// @param:
-// encoded_string: base64编码的数据的指针.
-// @return:
-// base64解码后的数据.
-std::string base64Decode(const std::string & encoded_string);
-
-// @param:
-// bytes_to_encode: 需要被编码的数据的指针.
-// @return:
-// urlencode后的数据.
-std::string urlEncode(const std::string& bytes_to_encode);
-
-// @param:
-// encoded_string: 需要被解码的数据的指针.
-// @return:
-// urldecode后的数据.
-std::string urlDecode(const std::string& encoded_string);
