+++ CppCurl/CppCurlTest.cpp
@@ -1,14 +1,18 @@
 #include "CppCurl.h"
+#include "Encoder.h"
 #include <iostream>
-
-char base64_img[] = "image=iVBORw0KGgoAAAANSUhEUgAAAfoAAACACAYAAADj26taAAAfAUlEQVR4nO2dCdxUVRmHj4agqLmgkpAmmQuKaG6logkaoAjmgiWWyqKCKGmkmQZurCGhomSkgCuEuCGpKFm5gxqCpqQobqAUGKRg2HKb5%2bC5nRlm5pv1u3fu%2fJ%2ffb2Dmm7l3zpx77ruc85733ShIYYQQQgiRSDaOugFCCCGEqB5S9EIIIUSCkaIXQgghEowUvRBCCJFgpOiFEEKIBCNFL4QQQiQYKXohhBAiwUjRCyGEEAlGil4IIYRIMFL0QgghRIKRohdCCCESjBS9EEIIkWCk6IUQQogEI0UvhBBCJBgpeiGEECLBSNELIYQQCUaKXgghhEgwUvRCCCFEgpGiF0IIIRKMFL0QQgiRYKTohRBCiAQjRS%2bEEEIkGCl6IYQQIsFI0QshhBAJRopeCCGESDBS9EIIIUSCkaIXQgghEowUvRBCCJFgpOiFEEKIBCNFL4QQQiQYKXohhBAiwUjRCyGEEAlGil4IIYRIMFL0QgghRIKRohdCCCESjBS9EEIIkWCk6IUQQogEI0UvhBBCJBgpeiGEECLBSNELIYQQCUaKXgghhEgwTaJuQBT8%2b9%2f%2fNitWrDDLli0zH3%2f8sWnZsqX50pe%2bZLbaaiuz0UYbRd28RPCf%2f%2fzHLF261PZxEAS2f3lsttlmUTdNCCHqirrz6P%2f%2b97%2bbgQMHmh133NEccMAB5sgjjzRt27Y122yzjRkzZoxZt25d1E2saVDqS5YsMT169DBf%2bcpXzCGHHGIOPfRQ89WvftW0a9fOzJgxw%2fz3v%2f%2bNuplCCFE31JWiX7NmjenUqZNZuXKluf%2f%2b%2b82%2f%2fvUvM3PmTHPKKafY93%2fyk5%2bYTTfd1Hz22WcRt7Q2QcnPmzfPKvWvf%2f3r5qmnnjLvvfeeue666%2bxsyVtvvWV69uxpjj32WNv3Qgghqs9GAdK5DsCLHD58uHn66afNfffdlzaFzHtXXHGFufrqq%2b3rrl27mgceeMA0bdo0qubWJKtXrzb777%2b%2f6dWrl%2b3PL3zhC%2bF7zz77rDn55JPtVD7ceOON5txzz42qqUIIUTfUjUePMp82bZr1KvHc%2fenjjTfe2PTp08dO58MjjzxiPX1RHLNnz7b9%2b%2bqrr1pF79uQTOH37t07fH3HHXfYdXwhhBDVpW4U%2fZ%2f%2f%2fGergN544w0zfvx4O6Xsw3ryN7%2f5zfA1nxWFg1J%2f88037fN7773XPPHEExso8u7du4fP%2f%2fSnPykeokAIHpVRFA%2fiEF9CGzQeoofrEIfxUAh1o%2bi%2f%2fOUvm1122cU%2b33777dOmlYFoe5S9g7XmWrmIcYGoegd9nMkee%2bwRPkfJL1iwoFHaVcsQLzJ06FAbNPr6669H3Zy65le%2f%2bpW9Dr%2f85S%2bt8RUFn376qV0CGzJkSCTfL9aDY3P66afba1ELMV11o%2bi33XZbc84555gjjjjCXHjhhaZVq1YbfGaLLbYIn3%2fjG99ozObVPBhKO%2b20kzn%2f%2fPPN8ccfb733Jk3Sd282b9487TXGl8gNXtuVV15pRo4cacftrrvuGnWT6pp%2b%2ffrZ7bjElsyfP7%2fRlT0K5fvf%2f75ZuHChVTAiOpB36BECjps1axb72cm6CcZzOC%2bddXkfhOoxxxxjHnvsMfv6ueeeqytlzzBAcDFg33%2f%2fffP222%2bHfYTn%2faMf%2fWiDWZBs5OpfmDVrVjh9v88%2b%2b5gXXnhBAY85YDziyY8YMcKcdNJJZurUqWaTTTaJulk1ixvff%2f3rX%2b2yHQqbvzHW2SHCoxDwqJm5%2bsc%2f%2fmGXAdlhkm2sV5p%2f%2fvOf9rs%2b%2bOADc%2ffdd0vRlwHXnV0%2fGE6Zso4lxYsuuqggWcd5fv3rX1sHktniRYsWWaUfSwJhmT59OgaPfZx99tlBahBE3aRGhd%2fbtm3bsA8yHymhVtb5U4oruOyyy8LzjRo1KkgZBRVqfbLgWkyZMsX2U7t27YJ33nkn6ibVPP7Yy3xceumlRY3FlBdnj0sp3mDp0qVVbPV61qxZE3Tr1s1%2b57Bhw%2by9JEqnkrKOa9G%2ff397XKdOnWKrN6ToU6Qs%2fSBl0duLdcghhwQrVqyIukmNDgM018Dff%2f%2f9yx7AKW8%2bPB9CK%2bWhVKjlyeO5554LWrVqZftq0qRJUTen5kGJH3744TnH929%2b85uiz3f99dfbY4899tjgk08%2bqVLLg2DdunXBxRdfbL%2friCOOiK0iqSUaknUpb7%2bo861evdrqDY7v06eP1Sdxo%2b4VPTftiBEj7EVq06ZNMG%2fevLq1mFG%2bn376afD444%2bnDf4BAwaU1SfcOM6QOv7446Xk80D%2fn3DCCaESKVboiOygMHm89NJLwZAhQ9LG9%2fLly4s%2bH8qiY8eO9vh%2b%2ffpVTbjPmTOnZINE5MbJuqlTp1ZE1vnnmTBhQhVaXB51r%2bidp7nDDjsECxcurFsl7zN37ty0wX%2fnnXeWfC4U1c9%2b9jN7Hqa45JHk54Ybbgj7%2fdFHH426OVUhynsMhdy9e%2fewj%2ffee%2b%2bSlTQCnXM0b948ePnllyvc0vX3zsknn2y%2f49RTT42lp1jrMHtWCVnHtdlvv%2f3sOTp06BA7OVfXin7RokXBXnvtFWy22WZaB%2fX4%2bc9%2fnjb433777ZLOw%2bAfOXKkPccll1wi77QBGIM777yz7a%2fvfe97iTQ6X3vtteA73%2flOZLM6COBtt902HNt9%2b%2fYt%2bVyrVq0KdtllF3uen%2f70pxUf3zNmzAjb%2bfvf%2f76i5xbrqZSsAyfreBDzFSfqZntdJhS36dy5s424XL58uUkJ2PA9Iu9T3lSErYuO1JgwTz75ZPi6TZs2WbciNgTR9w8%2f%2fLBJCUAblUp6Ybfdjvfo48cff7xi7a516HcSDb377rv2NVsUGyOau7FJGTO2zsQf%2fvCHSL6fFNgfffRR%2bJqCS6VC%2fQZqN8Bdd91la2hUCqLCU96lfc7WysMPP7xi5xbrySbrWrduXfL5Usa52Xrrre1zxkNUuRaykTxJUgDcROyr50F%2b9i233DJ8j4t%2f2223mXvuuSfCFkYH27r8wd%2bhQ4eiS%2feiyF9%2b%2bWW7lY49xxMmTEjbU88Wp%2bOOO07ZvTzY6kMNBkDgpLzeiFuUTNg261OuAkW4AwZMJdM6pzz4cDxQBKqQ7V6iODJl3WGHHVbW%2bdhi57Y9YsxiVMaFulP0KHlubnKvc9NnJnVBSZG%2btdB9tUnjlVdeMatWrQpfM%2fgz%2b6ghqBK433772b33VK7L9EzZb4pi82dR6h08XMYdHHXUUbaKoqgsCHaKKznI1OhnwyyFfffd1xZygunTp1dE0SODMBocRx99dNnnFBuSKetwaoqVdZkw%2b%2bJghi4u2VXL%2b1U1BsqFpDgtWrQwgwYNsoI101ulKAvTpySnqEd8QQgdO3Ys6njy3WMkUTgIhcX0vK%2fomTG59tpr7fN67eNM6JPf%2fva34eti%2b1wUBklufC%2bLaftykxDhaaPoSbRC2myUh1P8pcLswO23326fkz0SY0JUnnJlXTb8pSBXz8OvlBoVNanoEYxYzijpQqe0WC8577zzwnXhhx56KO%2fn61EJ0a9z584NX7Ne5eoDOMuUz6C4s03nI%2bhYsyTr2OjRo%2b0jF1i%2bmo5cD2P5xRdftM%2fx5CuV9cwvusE1S%2bKafzEwa%2bKvoyPYGcfB51nzeF5KPx188MHm5ptvts%2bRLxi6xS53%2bXAfOThXpe4TfqcbD%2fV%2b79EPmbKOJTP3HuSTdbngHG3btjWvvfaaTY%2bLcSlFXwTB52kLsZi5ESglS%2b7vAw44wCoXLHMEJp%2fLZgCw5kW6wkJxCq6eoP%2bef%2f758DUpgFHa48aNs6kily5daqc66fPvfve7NhjJ9TP9%2fuCDD4bBZA2Boi9HGCaJtWvXhoq%2bEgYQSouAs2eeecb88Y9%2fNEuWLLHTkig2p4TcFHO5U5W1hL8eS%2bplZvcIuqUkNUG5CGRSM9NHvFdo31DoxoGcYbawnNTOpIZ2YESUc584uclyGR4m7UN5sXTZq1cvW98jn9xMKijzTFmHUv7FL35h5Vw%2bWZcP%2bpZjUPSAMdGjR4%2bq%2fY5CqYm7nIFKxSiqR1E%2bdptttrGW0wMPPGDfZwpm1KhR5gc%2f%2bIENYKHww6233ppmmTOIL7744jDHceYDK8ynHpUQuw%2f88rwYUgg8bggi73lNn99yyy3mhhtuMNdff70Vcr6yv%2byyyxr8HvrWPYSx%2feuKYmBgltovCC%2bWngYPHmx3NZCX3XkpXDfOzbIJAo37he8h6NSvKlhtgohKa3DfY%2fg4KElNDAlT5ATlfu1rX7Pe%2fowZM%2bz7l156qbn88ssLUtj0MUqBKXcMq3LgGjqjD8qZWcTg416lLDdlupGbnI%2fzT5kyxV77a665xlx11VX2d2eTm0mlXFmXD2LAXIwFip7A48j7tJG38xUN2azOPffccH%2filVdeaf%2fGHmP25JKnmr%2b7xBI8tt5669glLKgFUkJvg5SQ7BNOGVh23zP7hOlXP51oyrCKutkh7NtnXDTGo5LJS%2fz9t2RpLPW3P%2fjgg8Huu%2b9uz3PooYfaDIdcL9677bbbghYtWgSbb7552vUdPXp0xX5HITzyyCP2e%2fm%2fMUkZQFlTnqacg2Dt2rW2jxjfY8eODd8j9Wwhe%2bM59qijjgqPmz17dsntJFvblltuGZ4rZZyUdB6uu5%2ff%2f7zzzrNyk78vW7YsSCl4%2b3eXsbLe5GY%2bWUc%2fOVlH2uFiZd2TTz4ZHsPxccgfEmuPHu%2fk7LPPNr%2f73e%2fsayxPPEZnHe255562jCfTUc4SB6YpI7egagw8CdaUfPbaay9rmWbuQOA6uMAV9sljsUY9BUz7hw8fHo6VakNJXryfcqc6U%2fegrabmcB54MfDb2d2AJw94bUSA%2b3uCme1iKyllhH38KOEkkxl4BcgOPHd%2f7P7whz%2b0OR%2bIxr7xxhvNKaecYqdi84Gs8ddh2T5aKszssFzmKKU0MbMX5K6YPHmyfX3qqafamRw3VnfccUdbYvX11183EydODI8rZSttLVKMrLviiitMp06d7PNCZd0OO%2bwQPv%2fb3%2f4W2SxWGusW3x%2bM7ts5OOTwzkHvkfcHb62J2vZYD1bQt7%2f97dAyIr1gruIRt9xyS5plhtcvioP%2bPuigg8I%2b3GKLLaxHmA2K%2fvj9PXPmzEZu7YbgVZFxzZgNvbZqPKh%2bVQmvnnP4s1GleLqk7fTblhLqWT%2bHt7jbbruFn2vWrFmje3BRefR4tH4fkXUuW%2bZBal%2f4Xm5KIRaUobBnz57hMcwKlArV8Nx5Nt1006K9QcbToEGDwnO0bNkyePXVV7N%2blrS9fp%2bkDPi6qCjJmC9U1qUUddGybuXKleHnmUWLQ22PJtf2vtpMX7te6S%2bccbU5rXkb88yg9mUZD%2bWSapeZNGlSWBseCHDZfPPNs34%2bM3Mb62%2biOAjY8oNThgwZknO7SWbGp3I8mEqBt4IXG5RgPQefByJlks%2b7IYCpEoFLeF%2bsFxbyndkgvoQ1RAfrg5leu4P2UkudOupAAFI9BF9lBpkSHEUyp2yzfowFP4EWXi8eYEMzhM2bNw%2bfOy%2buFO%2bYWRfHF7%2f4xaKO5Ttnz55t15Mdl1xyyQbxRw7Gnk%2b5gX%2b1AoGXhcq6zLwIXNuG4Lqxeyal4G3cB%2fIy6jr1TZauTf%2fDZ4uXmJWmvWkRTXssDHamzRwpz94GimSDwe1HqTKFFrd9yKUon1K%2bo5zlCj8iGb71rW%2fl%2fCx75X3KEWyVpNw90VF8F4q2EOGRDRQQgVb%2btDRKPtc4%2bOSTT9LuFVJAl7vEVezYdp8P1tfZKOrYUscXO0H8rVTt2rXLGWSXmeCESPVC2uk7IaVeTyCA0kGkdzG%2fGcVNcJ0DuUlmylywS8anEg5S0mUdTk1Dso73yNVC5D4wHnI5qY1Fk9YYop6yb9V%2bz0iVPDcaFunChQvDv%2bVLbMHn%2fWQjlRBelWbWrFlWGFdTERKvgDFUyndgtfqJRHbbbbe8ST8WL16c9joO%2b0RrmVI9%2bmXLlqUZxOTZZvzngq1kviLJ99lCYAvfmDFjSrrfELaZAjcfCNcLLrggbf2zUFzGQcdBBx2U93vcjAcUmukuc42%2bVMMXY8xRjEfP940dO9bGKznYQpfPoPFzibCtsBLKqDFkHbtEiDmplKzLF4PhjwUoVNb5ip7xEPV27SYXTB5izLBrzfT32piupw8wF57WeFttssGFyMwzT5BILvD%2b2VvvwIqN03Qk0zYUd2F7YDU54YQTzGmnnVbS4EdI%2bDnAyWiXrw%2f97T%2bAkYHgiFO%2f1wqMd1IGlwKKmyUXB9P2uZQD19g3nlnuIk1xObD9iG16pTBixIiijyEwrW%2ffvkUf53vzkC%2bnOUl1Pvjgg%2fA1QVpco4ZmcPz3fWOqWPzp9GKme5Ez06ZNS%2ftbvjz%2btHH%2b%2fPnha5Rdufcv%2fUR%2bE2pbVBNkXalODXIqU9blMlS5Z5jR8WEZpBBZ5xtN5MmImiZNdz3eXDw59Yi6JZ%2fDHl9%2ffyPTV%2fkigxmsvhXsJ6%2bIA0RoMkPh0r5Wk1JnMhYsWJDW56TczHUu1oR9AQFYxVLypcH42H777UPrv1AQqi4ph4OZr1zCDwVCoSEH91S5M19nnHGGNS6LgbgbIpfx%2fDDKi6GUnR2MV3%2f6de%2b99zbbbbddzs9nVtVjmr%2bQZRpfmJcy6%2bBo2bJl%2bBxZWCiZcpPZnXxyEwVG3zgqkcAKGYCsI8FWNSlnmRBj1%2b8njKFc9wH3TKmybsWKFeFz7u%2boid32OoSRH%2bzFAMx3o2FBOkh8sfvuu1e1faUQ9dazhiBrlg8eei7Y%2buNPubLNjO06onRQDE7RF7rGiZD2BRbk8%2bC4br4SO%2fDAA8s2zhCQxWaAc%2fcC%2f5eTPa5QUAhkdXSwDp1LSeCp%2bUl1gK2KhfST78WXo%2bj9Y%2f3AvIZgVtOXm4yFfHIH496HZDGVgL6Ks9GfKevyze5kk3W%2bIZYL7mE%2f1XI546FSxEoDcaP95S9%2fSfsbAzbXjYlX4weUILx864zzKQNbw%2fiCEPIpeqJ6fZj%2bbQyB3RBca2pAZ3q51YR92OUacYzNUi1%2b%2f7cy80UUfS7woP392URYO5xxkcT7hHHhj2%2fWrXPB5zJLi%2ba7F3z85ZdyPDiuo4vYxksv1PDLVGD5colkZt8jZsFvc5Llpj8WyFlB0aBc%2bE4kMNNZyHIKeon4FSDFMI%2boabLyhUnmqqE3m6c%2f2dIcdvowM7T3QaZFhAYZAUY%2b%2bSIiP%2fzwwzTDwPdSXnrpJZsMgzXE9u2j3S4Yd%2fxSjQgatmBlA6Hje4Uo%2bB%2f%2f%2bMexsOCZOqWWQWbgVbUg5enQoUMrruiLiVr2p%2fvzTcVzzkwF5iKsEUpnnXWWXf8mIUjcAlnLhd%2fnT4ETdJYLAtn8NXLiAQoJWKR%2fMz36UvvRjQeSKOHRFzoeMre45lvCpE%2f8oD2WUFx78fTZuUFisi5duhT%2fA2KOL%2btYlskluzB2%2fHTGyLqLLrqoIFnn77pgLMRBPjYZ13%2bCWS8CVpqnbxpgrmo121x3bDRx925bgk%2b%2buvB%2bEB7420MQ%2bgTB3XTTTZVtZALxDaF89bmJtmevuoOB75dljBKCX9hPjvHXGCAIKrHFDgHrrxn7wXX5QHj4a%2fv54ioyZ8r8%2ffNcUzKo9enTp9SfEGswxJhudTsbcuWOZ9rbz5bGNC170AsR0ijjSq3RIwNpr8uWSHbQQmoRZM485JObKHny8jv85QzqiTBdnTnLlxQKlXXEdZQq6zLX5xtjy2FDNPk44w9LltPI6BR95rQXgXYUY8iGKxzgcBcCixVrDGs8l3cq%2fg8FPRx%2bJatMMJ7cuiFTv9Scj0v8Ae3FW8vnscUVXyij6AsJNuJ9f20%2fX2QvpVP9mQ6unTMKXMrgWBTeqAKun5yid8WDMiE4zZcnCPZC089mevQNpczNB9eAyG6X76BQRZ8Zm8QyDTFL2cgsvOOWMzAIGSd8HxXbkogv64hzqYas82dXMMDiICM33jctqVxT02Gv4nNtVxKiGv29ipn7GB1MLfk3Jse5tRACaqjWhKJvzCQqtQpCwgXUYcln2zuMIGTPNNDPTNn7GcRE6fjR52TtKsQDwNP0l7X8fPk%2bLGGx%2f9zH5UjgexD6KLRu3bqV0vTYQz%2f5nltmDBAg8Nmuxbo4kN%2beRDOFrlH7W7YQ7OV49Ch6fzsxir4QMA78KPt8ctPf2ojcdPexk5vkxvcz%2fSUJX9Zxz1RD1vlGNdszY2FAB8vnBbece2RwwFH9g3GPvBWsq1Ku3UIhr%2fStt96almvazymespaCgQMH2vdOOumk8HOpDrU5jDmeyl2pmzt49913I%2fwltQN95qoA8qCylf%2ben0s9Zf0H8%2bbNq4uc2I2FXyVrn332KTi%2fObnrW7VqFebUXrPm%2f4UquG6TJk2y98W%2b%2b%2b6blq%2bbyl3gruuwYcMarcJWFLnulyxZEjRt2tR%2bb8pTtdXJHAsWLLB1C1zfkO%2b92L5IGUvh8WeccYa9LuXw4osvhuc7%2f%2fzzCzqG6%2f3www%2fnzNFPzvZscnO77bYLq4EiN3faaSdb3S6pZMq6m266Ke29cmUdn%2b3YsWN4jnvuuacaP6NoYlmmFsHnl33s3r27FQwUpkBoUZgDYbVq1aq0QUtZxgEDBtjnEyZMKKgYhVgPfd6jR4%2bwL%2fv37x%2fccccdQcrTCf%2fWu3fv4KOPPpKSrzCM08svvzzs55RXVdBxXIe5c%2beGx1Gilntg%2bPDhYUGos846K3j%2f%2fffTzo9RMWvWLPscg9g3EKpNFIqe%2fp0zZ074%2b%2bkbyvYi8Ckcw98wmO66666SDJ5Ro0aF5x43blzZ9weGws477xwqm0LlGPdw586dw7Z06dLFGnuDBw8O2rdvH2yyySbWkFm9enVw4oknhp%2fr16%2bfvd95PnXq1IqWYI4jGDaZsg4FXwlZR99ttdVW4XleeeWVKv2K4oilogcGLVXo2rVrl%2ba1o%2bwXL15sO5SLQGUgLFUElrthEXZxqBhUa9DnWLj0s%2btzLHy8lMmTJ9dNreooePTRR7POqDQESuCNN94IunXrZuuJu3NQyW%2fixInhNUOBIfSPO%2b648DPHHHOMreLWmERVvQ55gcFDvzjlzv8oRpTf888%2fX5JjwDG%2b0nziiScq0lZXDQ%2flTLXIQuE6jx8%2fPjj44IPDNiEbzzzzzOCxxx4LlTjyEVnq3%2bv8jrVr15bd%2flrAybo99tgj%2fP08xzCeMmVKybLONygxrsqd3akUsVX0gCJn4DL1xo2IJZbN2uRvvPfQQw%2fZKaqkW6TVxBlPb775pvUWGaiNNa1bz1CCGcGAgED5FOtJMP65Vs8%2b%2b6z1RBBUmefgNX9%2f7733bOlSfwq7sYhK0TsYyx9%2f%2fLH9fp675b5S4fjWrVvb39S1a9eKGcPMDDiFcd999xV9PL%2bN68wyQENy84UXXgjeeeedupObTtbhdS9atKgiss6f3RkzZkxsZj9jreiFqBcQsk5ING%2fePLHrpFEr%2bkrjxxPhIVbKKMZY23XXXe15%2b%2fbtGxuFIXKDwXjYYYfZa9amTZvgww8%2fjLpJITEIBxRCEB1O3niKzbBVLnMLVFIIvDK1tQ7R9m6vNZH9ROtXaisViat69epln1Od00%2bpKuIJ1QBdYiqq68Uhx71Dil6ImEDOByfcERp%2b7vKkcPTRR5s777zT%2fl%2frsB3Nlcg%2b88wzrXKuFGzJYssfxWlIApWZM0TECwzXqVOn2ucYff369YvHtrrPiU9LhKhz8AZR9OR%2buP32261CTBruN8YhiUg54M27srAkmCFtbKUFO8ld2NMO48ePTyufK%2bIF%2bQdmzJhhn%2ffv3z9vDv0okKIXIkaQyhalASNHjpRwjykI9rvvvts%2bHzBgwAapuysBhgOeYevWrW3inLFjxyZiySNpYPThzZN4iSJfgwYNil9BoGhDBIQQmRC57bZskRdCux7iBdejQ4cOYY6Pal%2bf6dOn2%2b9q1qxZ8NRTT1X1u0TxpAy%2bMCBz2rRpUTcnKxvxT6SWhhBiA8jJTu0G0nFSmOmcc86JuknCrPfeevbsae699167Jk%2blsmqn2SZWY%2bLEiWbgwIG2Kh0lhyniJKKHlNVdu3a1qZVHjx5t0%2bXGaW3eEb8WCSFs3WtyZlPQiYIaSY3CryVQ8tdcc41V8ieeeKItlNMYtTSIZ2AKnxz1lIkePHhwmJdfRAcljbk3UfLE1FBTIo5K3hL1lIIQIjck0mE%2fNRnM5s%2bfH3Vz6pqZM2fa6VmyzEWR8YwlAvbq0wbS9yrFd3SQ98LlzCeTZdyzhmrqXoiYs3DhQrNs2TLTpUuX%2bAX51BFUOqOsb8eOHSOrismswpw5c8yBBx6YswytaBxmz55t70e2isbWk%2f8cKXohhBAiwcTbDBFCCCFEWUjRCyGEEAlGil4IIYRIMFL0QgghRIKRohdCCCESjBS9EEIIkWCk6IUQQogEI0UvhBBCJBgpeiGEECLBSNELIYQQCUaKXgghhEgwUvRCCCFEgpGiF0IIIRKMFL0QQgiRYKTohRBCiAQjRS%2bEEEIkGCl6IYQQIsFI0QshhBAJRopeCCGESDBS9EIIIUSCkaIXQgghEowUvRBCCJFgpOiFEEKIBCNFL4QQQiQYKXohhBAiwUjRCyGEEAlGil4IIYRIMFL0QgghRIKRohdCCCESjBS9EEIIkWCk6IUQQogEI0UvhBBCJBgpeiGEECLBSNELIYQQCUaKXgghhEgwUvRCCCFEgvkf6g0k%2bL743dMAAAAASUVORK5CYII%3d";
-
+#include <fstream>
 
 int main()
 {
-	//std::string token = "24.2ab38b35c29d7c52e3ae47ce5c63b328.2592000.1596796408.282335-21194210";
-	//std::cout << formulaRecognition(base64_img, token) << std::endl;
-	std::string latex = "\\sigma(z)=\\frac{1}{1+e^{-z}}";
-	downloadRenderedFormula(latex,"latex_formula.gif","gif");
+
+	//std::string latex = "\\sigma(z)=\\frac{1}{1+e^{-z}}";
+	//downloadRenderedFormula(latex,"latex_formula.gif","gif");
+	//system("pause");
+
+	auto s = imageBase64UrlEncode("img.png");
+	s = "image=" + s;
+	std::string token = "24.2ab38b35c29d7c52e3ae47ce5c63b328.2592000.1596796408.282335-21194210";
+	std::cout << formulaRecognition(s, token) << std::endl;
 	system("pause");
 }
